mcpMode/mcpServer/requirements.txt
aiohttp>=3.8.0
typing-extensions>=4.0.0
python-json-logger>=2.0.0
pyyaml>=6.0.1
torch>=2.0.0
transformers>=4.30.0
